+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add Menu visualisation
+- Add icons to title actions
 - Multiline: updated the settings design, added chart groups
 - added legend highlight on the multiline component
 - added ability to bind the Y axis in the multiline component